add function refactor
